use-cases/anomaly-detection/backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
python-dotenv
plotly
scipy
matplotlib
scikit-learn
shap
joblib
pathlib2
sap-ai-sdk-gen
seaborn

